+++ src/index.ts
@@ -1,8 +1,8 @@
 import { Api, Bot, Context, InlineKeyboard } from 'grammy';
 import { AppUser, Channel } from './schema';
-import { v4 as uuidv4 } from 'uuid';
-import data from './wordList.json';
-
+import { getUUID, getUser, getChannel, getChannelFromJoin, 
+         userHasChannel, generateAlias } from './utility';
+import { sendBroadcast, replyToSender, sendToOwner } from './messageHandling';
 
 const bot = new Bot(process.env.BOT_TOKEN);
 
@@ -14,7 +14,7 @@ var chanList: Channel[] = [];
 bot.command('start', (ctx) => {
     // checks to see if the user exists and is already in the list
     // getUser returns -1 if the user does not exist
-    if (ctx.from && getUser(ctx.from.id) == -1) {
+    if (ctx.from && getUser(ctx.from.id, userList) == -1) {
         userList.push(<AppUser>{
             UUID: ctx.from.id,
             chatID: ctx.chat.id,
@@ -40,23 +40,12 @@ bot.command('start', (ctx) => {
     }
 });
 
-/* test method pls ignore
-bot.command('listchannels', (ctx) => {
-    ctx.reply("there are " + chanList.length + " channels");
-    for (let channel of chanList) {
-        ctx.reply(channel.owner.toString());
-    }
-    ctx.reply("done listing channels");
-    
-}); 
-*/ 
-
 // Create a new channel
 bot.command('newchannel', (ctx) => {
     if (ctx.from) {
-        if (getUser(ctx.from.id) > -1) {
+        if (getUser(ctx.from.id, userList) > -1) {
             ctx.reply("Creating a new channel...");
-            if (userHasChannel(ctx.from.id)) {
+            if (userHasChannel(ctx.from.id, chanList)) {
                 // true, user has a channel
                 //say no and tell them they have one
                 ctx.reply("user has channel");
@@ -74,7 +63,7 @@ bot.command('newchannel', (ctx) => {
                 let newChannelLength: number = chanList.push(newChannel);
 
                 // add the channel to the user's owned channels
-                let thisUser = userList[getUser(ctx.from.id)];
+                let thisUser = userList[getUser(ctx.from.id, userList)];
                 thisUser.channelsOwned.push(newChannel.UUID)
 
                 // set channel as the active channel
@@ -93,10 +82,10 @@ bot.command('newchannel', (ctx) => {
 // @TODO: Error validation - What happens if the user hasn't called /start?
 bot.command('joinchannel', (ctx) => {
     if (ctx.from) {
-        if (getUser(ctx.from.id) > -1) {
+        if (getUser(ctx.from.id, userList) > -1) {
             let joinCode: string = ctx.message.text.substring(13);
-            let channelInd = getChannelFromJoin(joinCode);
-            let userInd = getUser(ctx.from.id);
+            let channelInd = getChannelFromJoin(joinCode, chanList);
+            let userInd = getUser(ctx.from.id, userList);
             // if the number exists and is within the bounds of normalcy
             if (channelInd != null && channelInd >= 0 && channelInd < chanList.length) {
                 // add channel to user's default list
@@ -110,7 +99,9 @@ bot.command('joinchannel', (ctx) => {
                 chanList[channelInd].senderAlias.set(alias, ctx.from.id);
                 chanList[channelInd].senderAliasReverse.set(ctx.from.id, alias);
 
-                ctx.reply("Alrighty, " + ctx.from.first_name + ", you've joined a channel that forwards to:" + userList[getUser(chanList[channelInd].owner)].nameOnMsg);
+                ctx.reply("Alrighty, " + ctx.from.first_name + 
+                          ", you've joined a channel that forwards to:" + 
+                          userList[getUser(chanList[channelInd].owner, userList)].nameOnMsg);
             } else {
                 ctx.reply("Oops, looks like that's not a valid code! Try again.");
             }
@@ -134,44 +125,29 @@ bot.command('managechannel', (ctx) => {
     // NYE
 });
 
-// bot.command('messagechannel', (ctx) => {
-//     // list channels via inline bot (index them based on where they are in user's chatSender)
-//     if (!ctx.from) {
-//         return; // this is stupid
-//     }
-//     const InlineChannelMenu = new InlineKeyboard();
-
-//     let currentUser = userList[getUser(ctx.from.id)];
-
-//     for (let i = 0; i < currentUser.chatSender.length; i++) {
-//         InlineChannelMenu.text( (i+1) + ": " + getOwnerNameFromChannel(currentUser.chatSender[i]), "");
-//     }
-// })
-
-
 // deal with non-command user text messages
 bot.on('message', (ctx) => {
-    if (getUser(ctx.from.id) > -1) {
+    if (getUser(ctx.from.id, userList) > -1) {
         // if user is in a channel
-        if(userList[getUser(ctx.from.id)].activeChannel){
+        if(userList[getUser(ctx.from.id, userList)].activeChannel){
 
-            let myUser = userList[getUser(ctx.from.id)];
-            let myChannel = chanList[getChannel(myUser.activeChannel)];
+            let myUser = userList[getUser(ctx.from.id, userList)];
+            let myChannel = chanList[getChannel(myUser.activeChannel, chanList)];
             
             // check if user owns channel
             if (myChannel.owner == ctx.from.id){
                 // check if message is a broadcast or reply
                 if (ctx.message.reply_to_message){
                     // message is a reply
-                    replyToSender(ctx, userList[getUser(ctx.from.id)], myChannel);
+                    replyToSender(ctx, userList[getUser(ctx.from.id, userList)], myChannel, userList, chanList);
                 } else {
                     // message is a broadcast
-                    sendBroadcast(ctx, userList[getUser(ctx.from.id)], myChannel);
+                    sendBroadcast(ctx, userList[getUser(ctx.from.id, userList)], myChannel, userList);
                 }
             // User has send access to channel
             } else if (myChannel.senders.includes(ctx.from.id)){
                 // Send message to channel owner
-               sendToOwner(ctx, myUser, myChannel);
+               sendToOwner(ctx, myUser, myChannel, userList);
 
             // user is not in a channel
             } else {
@@ -187,149 +163,3 @@ bot.start();
 // Enable graceful stop
 process.once('SIGINT', () => bot.stop());
 process.once('SIGTERM', () => bot.stop());
-
-////////////////////////////////////////////////////
-//
-//              utility functions
-//
-/// functions to support above methods and help 
-///        make things more readable 
-//
-////////////////////////////////////////////////////
-
-// Generates a UUID from the UUID package
-function getUUID(): string {
-    return uuidv4();
-}
-
-// Returns the index of the user that has the UUID specified
-function getUser(UUID: number): number {
-    return userList.findIndex(AppUser => AppUser.UUID === UUID);
-}
-
-// Returns the telegram user ID of a sender based on their alias
-function getIDfromSender(user: string, channelID: string): number | undefined {
-    if (chanList[getChannel(channelID)].senderAlias.get(user)) {
-        return chanList[getChannel(channelID)].senderAlias.get(user);
-    }
-    return undefined;
-}
-
-// Returns the index of the channel that has the UUID specified
-function getChannel(UUID: string): number {
-    return chanList.findIndex(Channel => Channel.UUID === UUID);
-}
-
-// Returns the index of the channel the user is trying to join
-function getChannelFromJoin(joinCode: string): number {
-    return chanList.findIndex(Channel => Channel.joinLink === joinCode);
-}
-
-// Returns the username extracted from a user's message
-function getUserFromMessage(message: string): string | null {
-    let re = new RegExp('(?<=<).*(?=>)');
-    let resultArray = re.exec(message);
-    if (resultArray) {
-        let result = resultArray[0];
-        if (process.env.NODE_ENV == 'dev'){
-            console.log(result);
-        }
-        return result;
-    }
-    return null;
-}
-
-// Returns true if the message is a forwarded user message, false if bot message
-function isUserMessage(message: string): boolean {
-    let re = new RegExp("^<");
-    return message.search(re) >= -1;
-}
-
-// 
-function getOwnerNameFromChannel(UUID: string): string {
-    // get the channel
-    let channelInd = getChannel(UUID);
-    // get the owner
-    let ownerInd = getUser(chanList[channelInd].owner);
-    // get the owner's name and return
-    return userList[ownerInd].nameOnMsg;
-}
-
-// Returns true if the channel exists and else otherwise
-function userHasChannel(UUID: number): boolean {
-    let channel = chanList.findIndex(Channel => Channel.owner === UUID);
-    if (channel != -1) {
-        // if the channel IS NOT undefined (exists)
-        return true;
-    } else {
-        // if the channel IS undefined
-        return false;
-    }
-}
-
-function generateAlias(): string {
-    let alias = "";
-
-    let adjList = data.adj;
-    let nounList = data.noun;
-
-    alias += adjList[Math.floor(Math.random() * adjList.length)];
-    alias += nounList[Math.floor(Math.random() * nounList.length)];
-
-    return alias;
-}
-
-function sendBroadcast(ctx: Context, owner: AppUser, channel: Channel) {
-    if (ctx.message){
-        for (let senderID of channel.senders) {
-            let user = userList[getUser(senderID)];
-            if (ctx.message.text) {
-                let username = owner.nameOnMsg;
-                let wrappedMessage = "<" + username + "> " + ctx.message.text;
-                ctx.api.sendMessage(user.chatID, wrappedMessage);
-            }
-        }
-    } else {
-        ctx.reply("Something went wrong. Please try again.");
-    }
-}
-
-function replyToSender(ctx: Context, owner: AppUser, channel: Channel) {
-    if (ctx.message && ctx.message.reply_to_message && ctx.from ){
-        if (ctx.message.reply_to_message.text && isUserMessage(ctx.message.reply_to_message.text)) {
-            let sender = getUserFromMessage(ctx.message.reply_to_message.text);
-            if (sender) {
-                let senderID = getIDfromSender(sender, userList[getUser(ctx.from.id)].activeChannel);
-                if (senderID) {
-                    console.log("Sender is " + sender + " with ID " + senderID)
-                    let user = userList[getUser(senderID)];
-                    let owner = userList[getUser(channel.owner)];
-                    let username = owner.nameOnMsg;
-                    let wrappedMessage = "_<" + username + "\\>_ " + ctx.message.text;
-                    ctx.api.sendMessage(user.chatID, wrappedMessage, {
-                        parse_mode: "MarkdownV2",
-                    });
-                }
-            }
-        // owner is trying to reply to the bot
-        } else if (ctx.message.text) {
-            ctx.reply("I'm sorry, messages cannot be broadcast as replies.");
-        }
-    }
-}
-
-function sendToOwner(ctx: Context, sender: AppUser, channel: Channel) {
-    let owner = userList[getUser(channel.owner)];
-    if (ctx.message && ctx.message.text && ctx.from) {
-        if (process.env.NODE_ENV == 'dev'){
-            console.log(JSON.stringify(channel));
-        }
-        let alias = channel.senderAliasReverse.get(sender.UUID);
-        let wrappedMessage = "<" + alias + "> " + ctx.message.text;
-        ctx.api.sendMessage(owner.chatID, wrappedMessage, {
-            parse_mode: "MarkdownV2",
-        });
-    } else {
-        ctx.reply("Something went wrong, please try again.");
-    }   
-}
